+++ client/v2/autocli/flags.go
@@ -1,3 +1,0 @@
-package autocli
-
-var flagNoIndent = "no-indent"