+++ x/protocolpool/testutil/app_config.go
@@ -12,7 +12,6 @@ import (
 	_ "github.com/cosmos/cosmos-sdk/x/genutil"        // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/gov"            // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/mint"           // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/params"         // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/staking"        // import as blank for app wiring
 )
 
@@ -22,7 +21,6 @@ var AppConfig = configurator.NewAppConfig(
 	configurator.StakingModule(),
 	configurator.TxModule(),
 	configurator.ConsensusModule(),
-	configurator.ParamsModule(),
 	configurator.GenutilModule(),
 	configurator.DistributionModule(),
 	configurator.MintModule(),